feat/fix: Criação e ajuste de DTOs juntamente com a cricação e ajuste de novos entitys

+++ Projeto_Spring_MicroS_Clima/src/main/java/br/com/orati/cepclima/model/Clima.java
@@ -23,5 +23,5 @@ public class Clima {
         Float temperatura;
         Float sensacaoTermica;
         Integer condicaoClimatica;
-        Integer velocidadeVento;
+        Float velocidadeVento;
 }
